Added functionnal UI (CPU Page), Custom CPU renaming and inspecting
Hey ! The first commit was not that voluntary, you know... So, here's one that works.

- Add: Buttons to write CPU name value
-Add: Text boxes where to type custom CPU name
-Add: A stunning button to inspect and display the actual CPU name stored in Registry.
-Add: Formatting system for Powershell
-Add: Comments all over the code so it's more readable for everyone learning how to code or just the curious ones

-Fixes: Writing and reading registry keys has been optimized and is now working perfectly.

Feel free to test it, there will be more updates about the UI but at the moment it's working and usable !

IMPORTANT: Don't forget to give the program administrator privileges or it probably won't work.